requirements.txt
pandas==1.5.3
yfinance==0.2.52
python-dotenv==0.21.1
openai
langchain
langchain_core>=0.2.38
langchain_openai
langgraph
numpy